add error code to 404, add description to posts
This change adds specific error code to default 404 error page to
make the error page clearer, especially if other pages are added
for more specific errors.

The change also adds the description of a blog post as an h3 tag
below the post's title. Before, descriptions of posts were only
found on the main page.